Graphing
This is the code that will graph our data. It will take in two column of data, each column with a title at position 0. The input needed for the data must be tab separated (or consistently one separation value) text file. The data can be any type, but it will always be treated as strings.