Semi_Final_Data_Grapher
This is the (nearly) completed version of the data grapher. It takes on a tab separated text file with the first row as the column titles, and the other rest as data. It only takes in two columns of data, where the second column must be numerical to plot properly. However, I believe strings would work on the second column, but I've yet to test it.